requirements.txt
airtest==1.0.27
asn1crypto==0.24.0
beautifulsoup4==4.8.0
blinker==1.4
brotlipy==0.7.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==6.7
comtypes==1.1.7
cryptography==2.3.1
decorator==4.4.0
facebook-wda==0.3.4
h11==0.7.0
h2==3.1.0
hpack==3.0.0
hrpc==1.0.8
html5lib==1.0.1
hyperframe==5.2.0
idna==2.8
Jinja2==2.10.1
kaitaistruct==0.8
ldap3==2.5.2
MarkupSafe==1.1.1
mitmproxy==4.0.4
mss==4.0.3
numpy==1.16.4
opencv-contrib-python==3.4.2.17
passlib==1.7.1
Pillow==6.1.0
pocoui==1.0.76
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
pydivert==2.1.0
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
pypiwin32==223
pywin32==224
pywinauto==0.6.3
requests==2.22.0
retry==0.9.2
ruamel.yaml==0.15.100
six==1.12.0
sortedcontainers==2.0.5
soupsieve==1.9.2
tornado==5.1.1
urllib3==1.25.3
urwid==2.0.1
webencodings==0.5.1
websocket-client==0.56.0
wsproto==0.11.0
